new geojson

+++ src/main/java/com/giant/demo/preprocessing/PreProcess.java
@@ -52,11 +52,11 @@ public class PreProcess {
             Integer votes = (int) (double) p.get("votes");
             Double demo = (double) p.get("demo");
             Double rep = (double) p.get("rep");
-            double africanAmerican = (long) p.get("africanAmerican") / (double)pop;
-            double asian = (long) p.get("asian") / (double)pop;
-            double latinAmerican = (long) p.get("latinAmerican") / (double)pop;
-            double white = (long) p.get("white") / (double)pop;
-            double other = (long) p.get("other") / (double)pop;
+            double africanAmerican = (double)p.get("africanAmerican") / (double)pop;
+            double asian = (double) p.get("asian") / (double)pop;
+            double latinAmerican = (double) p.get("latinAmerican") / (double)pop;
+            double white = (double) p.get("white") / (double)pop;
+            double other = (double) p.get("other") / (double)pop;
             Demographics demographics = new Demographics(africanAmerican, asian, latinAmerican, white, other, pop);
 
             JSONArray array = (JSONArray) p.get("neighbor");
